requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.1
PyGithub==1.58.1
Requests==2.31.0
scikit_learn==1.2.2